+++ akshare/__init__.py
@@ -2377,9 +2377,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.9.59 fix: fix stock_hot_search_baidu interface
 1.9.60 add: add stock_hk_fhpx_detail_ths interface
 1.9.61 fix: fix stock_lhb_detail_daily_sina interface
+1.9.62 fix: fix bond_zh_us_rate interface
 """
 
-__version__ = "1.9.61"
+__version__ = "1.9.62"
 __author__ = "AKFamily"
 
 import sys