+++ akshare/bond/bond_em.py
@@ -1,19 +1,19 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2022/3/5 14:18
+Date: 2022/4/23 17:18
 Desc: 东方财富网-数据中心-经济数据-中美国债收益率
-http://data.eastmoney.com/cjsj/zmgzsyl.html
+https://data.eastmoney.com/cjsj/zmgzsyl.html
 """
 import pandas as pd
 import requests
 from tqdm import tqdm
 
 
-def bond_zh_us_rate() -> pd.DataFrame:
+def bond_zh_us_rate(start_date: str = "19901219") -> pd.DataFrame:
     """
     东方财富网-数据中心-经济数据-中美国债收益率
-    http://data.eastmoney.com/cjsj/zmgzsyl.html
+    https://data.eastmoney.com/cjsj/zmgzsyl.html
     :return: 中美国债收益率
     :rtype: pandas.DataFrame
     """
@@ -51,6 +51,12 @@ def bond_zh_us_rate() -> pd.DataFrame:
         data_json = r.json()
         temp_df = pd.DataFrame(data_json["result"]["data"])
         big_df = pd.concat([big_df, temp_df], ignore_index=True)
+        temp_date_list = pd.to_datetime(big_df["SOLAR_DATE"]).dt.date.to_list()
+        if pd.to_datetime(start_date) in pd.date_range(
+            temp_date_list[-1], temp_date_list[0]
+        ):
+            break
+
     big_df.rename(
         columns={
             "SOLAR_DATE": "日期",
@@ -86,24 +92,27 @@ def bond_zh_us_rate() -> pd.DataFrame:
             "美国GDP年增率",
         ]
     ]
-    big_df["日期"] = pd.to_datetime(big_df["日期"]).dt.date
-    big_df["中国国债收益率2年"] = pd.to_numeric(big_df["中国国债收益率2年"])
-    big_df["中国国债收益率5年"] = pd.to_numeric(big_df["中国国债收益率5年"])
-    big_df["中国国债收益率10年"] = pd.to_numeric(big_df["中国国债收益率10年"])
-    big_df["中国国债收益率30年"] = pd.to_numeric(big_df["中国国债收益率30年"])
-    big_df["中国国债收益率10年-2年"] = pd.to_numeric(big_df["中国国债收益率10年-2年"])
-    big_df["中国GDP年增率"] = pd.to_numeric(big_df["中国GDP年增率"])
-    big_df["美国国债收益率2年"] = pd.to_numeric(big_df["美国国债收益率2年"])
-    big_df["美国国债收益率5年"] = pd.to_numeric(big_df["美国国债收益率5年"])
-    big_df["美国国债收益率10年"] = pd.to_numeric(big_df["美国国债收益率10年"])
-    big_df["美国国债收益率30年"] = pd.to_numeric(big_df["美国国债收益率30年"])
-    big_df["美国国债收益率10年-2年"] = pd.to_numeric(big_df["美国国债收益率10年-2年"])
-    big_df["美国GDP年增率"] = pd.to_numeric(big_df["美国GDP年增率"])
+    big_df["日期"] = pd.to_datetime(big_df["日期"])
+    big_df["中国国债收益率2年"] = pd.to_numeric(big_df["中国国债收益率2年"], errors="coerce")
+    big_df["中国国债收益率5年"] = pd.to_numeric(big_df["中国国债收益率5年"], errors="coerce")
+    big_df["中国国债收益率10年"] = pd.to_numeric(big_df["中国国债收益率10年"], errors="coerce")
+    big_df["中国国债收益率30年"] = pd.to_numeric(big_df["中国国债收益率30年"], errors="coerce")
+    big_df["中国国债收益率10年-2年"] = pd.to_numeric(big_df["中国国债收益率10年-2年"], errors="coerce")
+    big_df["中国GDP年增率"] = pd.to_numeric(big_df["中国GDP年增率"], errors="coerce")
+    big_df["美国国债收益率2年"] = pd.to_numeric(big_df["美国国债收益率2年"], errors="coerce")
+    big_df["美国国债收益率5年"] = pd.to_numeric(big_df["美国国债收益率5年"], errors="coerce")
+    big_df["美国国债收益率10年"] = pd.to_numeric(big_df["美国国债收益率10年"], errors="coerce")
+    big_df["美国国债收益率30年"] = pd.to_numeric(big_df["美国国债收益率30年"], errors="coerce")
+    big_df["美国国债收益率10年-2年"] = pd.to_numeric(big_df["美国国债收益率10年-2年"], errors="coerce")
+    big_df["美国GDP年增率"] = pd.to_numeric(big_df["美国GDP年增率"], errors="coerce")
     big_df.sort_values("日期", inplace=True)
-    big_df.reset_index(inplace=True, drop=True)
+    big_df.set_index(["日期"], inplace=True)
+    big_df = big_df[pd.to_datetime(start_date) :]
+    big_df.reset_index(inplace=True)
+    big_df["日期"] = pd.to_datetime(big_df["日期"]).dt.date
     return big_df
 
 
 if __name__ == "__main__":
-    bond_zh_us_rate_df = bond_zh_us_rate()
+    bond_zh_us_rate_df = bond_zh_us_rate(start_date="19901219")
     print(bond_zh_us_rate_df)