+++ docs/changelog.md
@@ -56,6 +56,10 @@
 
 ## 更新说明
 
+1.9.62 fix: fix bond_zh_us_rate interface
+
+    1. 修复 bond_zh_us_rate 接口
+
 1.9.61 fix: fix stock_lhb_detail_daily_sina interface
 
     1. 修复 stock_lhb_detail_daily_sina 接口
@@ -2222,6 +2226,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.62 fix: fix bond_zh_us_rate interface
 
 1.9.61 fix: fix stock_lhb_detail_daily_sina interface